Use github API to determine default branch
Context
---

In newer repositories, the default branch is now main instead of master.

Problem
---

In the current implementation, maiao is not aware of the github settings
and is restricted to the local ones, that may mismatch.
This causes a `reference not found` error.

Solution
---

Update the pull requester interface to allow retrieving the repository
settings and determine the relevant default branch

Change-Id: I3298d98682d6fe6e4ebc226c40e0abce7cac2a26