Fix Commit message parser
With the current (naive) commit message parser, any line that includes a
colon (:) will basically be consider a HEADER. That becomes a problem if
you try to include an HTTP link in your commit message. It will be
consider a HEADER and the removed from the final Pull Request message.

For example, if your commit message was:
`
Fixed critical security vulnerability

Vulnerability reported [here](http://www.maiao.com)

As you can read in the previous link it is really severe
`

Maiao will craft a PR that will read:
`
Fixed critical security vulnerability

As you can read in the previous link it is really severe
`

Leaving a broken message

Proposed solution
----------------
It is kinda of naive as well. It assumes that, in a header, the colon : will
NEVER be followed by an slash /