Better support for Github Enterprise
Background
=====

People may use Maiao for [Github Enterprise][ghe] and github.com.

Problem
=====

Currently, the burden of changing the `GITHUB_TOKEN` environment variable
is on the user. If the user doesn't want to set a `.netrc` file for
different hosts, they need to things such as
`export GITHUB_TOKEN=$GITHUB_ENTERPRISE_TOKEN`, which is not very friendly.

Goal
=====

To support domains out of `github.com` through the `GITHUB_ENTERPRISE_TOKEN`
environment variable.

Caveats
=====

To avoid breaking changes with the current API, there's a fallback for
domains other than `github.com` to search for the `GITHUB_TOKEN`.

[ghe]: https://github.com/enterprise

Change-Id: Ib5874575a3c7ced79748288582249571d7771c05