Add support for system and pass keyrings
In the current implementation, maiao uses unsecure storage for github
credentials. Either ~/.netrc or GITHUB_TOKEN environment variable.

This causes a security issue as the secrets are not protected and
available from anybody on the computer.

The goal of this commit is to rely on safer credential stores like
[pass], or the system keychain to store credentials.

Those credential stores uses non-trivial format and hence, implement a
logic to retrieve login details to the user prior to storing them in the
credentials store.

[pass]: https://www.passwordstore.org

Change-Id: I4c34e4e45bce64b77cc0a3adbeebdc7e9422f1e3