Add pre-commit-message hook
In many cases, our GitHub users have PullRequest templates in their
repositories to guide users toward creating better quality PRs.

When using maiao, this behaviour is lost and the pull request templates
are not used.

With maiao's philosophy that every single (small) commit turns into a
Pull Request, this commit implements an optional and experimental hook
to insert the pull request template.

This injestion is skept when using `git commit -m "my-message"` as git
won't launch any editor with interactions with the user.

Change-Id: I171787936ff8a097bd041a3bd3542207ded7a91a